@@ -2,7 +2,7 @@ package com.twiceyuan.imagepickcompat.utils;
 
 import android.content.Context;
 import android.net.Uri;
-import android.support.v4.content.FileProvider;
+import androidx.core.content.FileProvider;
 
 import com.twiceyuan.imagepickcompat.Constants;
 
